added Auth to Journal Entry

+++ src/main/java/com/adityarastogi/lumiJournal/service/JournalEntryService.java
@@ -37,7 +37,7 @@ public class JournalEntryService {
 
             // user.setUserName(null);  //just to check transaction
 
-            userService.saveEntry(user);   
+            userService.saveUser(user);   
         } catch (Exception e) {
             // log.error("Error saving journal entry", e);
             System.out.println(e);
@@ -57,10 +57,21 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            
+            if(removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleteing the Journal entry", e);
+        }
+        return removed;
     }
 }
